[llvm][offload] Move AMDGPU offload utilities to LLVM
This patch moves utilities from `offload/plugins-nextgen/amdgpu/utils/UtilitiesRTL.h`
to `llvm/Frontend/Offloading/AMDGPU/ObjectUtilities.h` to be reused by other
projects.

Concretely the following changes were made:
- Rename `KernelMetaDataTy` to `AMDGPUKernelMetaData`.
- Remove unused fields `KernelObject`, `KernelSegmentSize`, `ExplicitArgumentCount`
  and `ImplicitArgumentCount` from `AMDGPUKernelMetaData`.
- Return the produced error if `ELFObj.sections()` failed instead of using `cantFail`.
- Added `AGPRCount` field to `AMDGPUKernelMetaData`.
