+++ llvm/lib/Frontend/Offloading/AMDGPU/CMakeLists.txt
@@ -1,0 +1,8 @@
+add_llvm_component_library(LLVMFrontendOffloadingAMDGPU
+  ObjectUtilities.cpp
+
+  LINK_COMPONENTS
+  Support
+  BinaryFormat
+  Object
+)
